remove helpdesk button from homepage (#7)
At last month's meeting we decided to take a bit of a break over the summer, and will be discussing the future of the data helpdesk afterwards. In the meantime we are removing the "ask helpdesk" button from top landing page of the website (while leaving the one lower down with the full description.